Fixed something in GUI

+++ gui/my_gui.py
@@ -47,9 +47,9 @@ class MyGUI:
         self.canvas.delete("all")  # Clear the canvas before drawing
 
         for cup in cup_positions:
-            x = cup[0] * 2.06
-            y = cup[1] * 2.06
-            radius = cup[2] * 2.06
+            x = int(cup[0] * 2.06)
+            y = int(cup[1] * 2.06)
+            radius = int(cup[2] * 2.06)
 
             print(f"Comparing cup at position: {x, y, radius} with hit_cups: {self.hit_cups}")
